move platformservice to a providedInRoot directory
