+++ projects/laji/src/app/+project-form/form/named-place/np-print/np-print.component.ts
@@ -8,7 +8,7 @@ import { TranslateService } from '@ngx-translate/core';
 import { UserService } from '../../../../shared/service/user.service';
 import { FooterService } from '../../../../shared/service/footer.service';
 import { LajiApi, LajiApiService } from '../../../../shared/service/laji-api.service';
-import { PlatformService } from '../../../../shared-modules/platform/platform.service';
+import { PlatformService } from '../../../../root/platform.service';
 import { NamedPlacesService } from '../../../../shared/service/named-places.service';
 import { ProjectFormService } from '../../../project-form.service';
