+++ projects/laji/src/app/+viewer/viewer-print/viewer-print.component.ts
@@ -7,7 +7,7 @@ import { Subscription } from 'rxjs';
 import { ActivatedRoute } from '@angular/router';
 import { LajiApi, LajiApiService } from '../../shared/service/laji-api.service';
 import * as FileSaver from 'file-saver';
-import { PlatformService } from '../../shared-modules/platform/platform.service';
+import { PlatformService } from '../../root/platform.service';
 
 @Component({
   selector: 'laji-viewer-print',
