+++ projects/laji/src/app/locale/locale-en.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { LocaleComponent } from './locale.component';
 import * as moment from 'moment';
-import { PlatformService } from '../shared-modules/platform/platform.service';
+import { PlatformService } from '../root/platform.service';
 
 @Component({
   selector: 'laji-locale-en',
