+++ projects/laji/src/app/shared-modules/app-component/app.component.ts
@@ -7,7 +7,7 @@ import { Global } from '../../../environments/global';
 import { setTheme } from 'ngx-bootstrap/utils';
 import { RouteDataService } from '../../shared/service/route-data.service';
 import { HeaderService } from '../../shared/service/header.service';
-import { PlatformService } from '../platform/platform.service';
+import { PlatformService } from '../../root/platform.service';
 import { HistoryService } from '../../shared/service/history.service';
 import { Util } from '../../shared/service/util.service';
