+++ projects/laji/src/app/shared-modules/audio-viewer/directive/requires-audio-support.directive.ts
@@ -1,5 +1,5 @@
 import { ComponentFactoryResolver, Directive, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
-import { PlatformService } from '../../platform/platform.service';
+import { PlatformService } from '../../../root/platform.service';
 import { AudioNotSupportedErrorComponent } from './audio-not-supported-error.component';
 import { AudioIosWarningComponent } from './audio-ios-warning.component';
