+++ projects/laji/src/app/shared-modules/datatable/datatable/datatable.component.ts
@@ -8,7 +8,7 @@ import { DatatableTemplatesComponent } from '../datatable-templates/datatable-te
 import { Logger } from '../../../shared/logger/logger.service';
 import { FilterByType, FilterService } from '../../../shared/service/filter.service';
 import { LocalStorage } from 'ngx-webstorage';
-import { PlatformService } from '../../platform/platform.service';
+import { PlatformService } from '../../../root/platform.service';
 import { DatatableUtil } from '../service/datatable-util.service';
 import { Util } from '../../../shared/service/util.service';
