+++ projects/laji/src/app/shared-modules/info/info/info.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, HostListener, Input, ViewChild, OnInit } from '@angular/core';
 import { ModalDirective } from 'ngx-bootstrap/modal';
 import { PopoverDirective } from 'ngx-bootstrap/popover';
-import { PlatformService } from '../../platform/platform.service';
+import { PlatformService } from '../../../root/platform.service';
 
 @Component({
   selector: 'laji-info',
