+++ projects/laji/src/app/shared-modules/utilities/directive/dumb-directives/browser-only.directive.ts
@@ -1,5 +1,5 @@
 import { Directive, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
-import { PlatformService } from '../../../platform/platform.service';
+import { PlatformService } from '../../../../root/platform.service';
 
 @Directive({ selector: '[lajiBrowserOnly]' })
 export class BrowserOnlyDirective implements OnInit {
