+++ projects/laji/src/app/shared-modules/utilities/directive/dumb-directives/server-only.directive.ts
@@ -1,5 +1,5 @@
 import { Directive, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
-import { PlatformService } from '../../../platform/platform.service';
+import { PlatformService } from '../../../../root/platform.service';
 
 @Directive({ selector: '[lajiServerOnly]' })
 export class ServerOnlyDirective implements OnInit {
