+++ projects/laji/src/app/shared/directive/hide-for-ie.directive.ts
@@ -1,5 +1,5 @@
 import { Directive, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
-import { PlatformService } from '../../shared-modules/platform/platform.service';
+import { PlatformService } from '../../root/platform.service';
 
 @Directive({ selector: '[lajiHideForIe]' })
 export class HideForIeDirective implements OnInit {
