+++ projects/laji/src/app/shared/directive/lazy-image.directive.ts
@@ -8,7 +8,7 @@ import {
   NgZone,
   OnDestroy
 } from '@angular/core';
-import { PlatformService } from '../../shared-modules/platform/platform.service';
+import { PlatformService } from '../../root/platform.service';
 
 @Directive({
   selector: 'img[lajiLazyImage]'
