+++ projects/laji/src/app/shared/interceptor/timeout.interceptor.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { PlatformService } from '../../shared-modules/platform/platform.service';
+import { PlatformService } from '../../root/platform.service';
 import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
 import { EMPTY, Observable, throwError } from 'rxjs';
 import { catchError, timeout } from 'rxjs/operators';
