+++ projects/laji/src/app/shared/service/form-permission.service.ts
@@ -8,7 +8,7 @@ import { UserService } from './user.service';
 import { catchError, map, switchMap, take, tap } from 'rxjs/operators';
 import { FormService } from './form.service';
 import { TranslateService } from '@ngx-translate/core';
-import { PlatformService } from '../../shared-modules/platform/platform.service';
+import { PlatformService } from '../../root/platform.service';
 import RestrictAccess = Form.RestrictAccess;
 
 export interface Rights {
